+++ layout_evaluation/src/metrics/bigram_metrics/oxey_sfbs.rs
@@ -42,10 +42,10 @@ impl BigramMetric for OxeySfbs {
         &self,
         k1: &LayerKey,
         k2: &LayerKey,
-        weight: f64,
-        _total_weight: f64,
+        weight: f32,
+        _total_weight: f32,
         _layout: &Layout,
-    ) -> Option<f64> {
+    ) -> Option<f32> {
         if self.exclude_modifiers && (k1.is_modifier.is_some() || k2.is_modifier.is_some()) {
             return Some(0.0);
         }
