+++ layout_evaluation/src/metrics/bigram_metrics/symmetric_handswitches.rs
@@ -35,10 +35,10 @@ impl BigramMetric for SymmetricHandswitches {
         &self,
         k1: &LayerKey,
         k2: &LayerKey,
-        weight: f64,
-        _total_weight: f64,
+        weight: f32,
+        _total_weight: f32,
         _layout: &Layout,
-    ) -> Option<f64> {
+    ) -> Option<f32> {
         if k1.key.hand != k2.key.hand
             && k1.key.symmetry_index == k2.key.symmetry_index
             && k1.key.finger != Finger::Thumb
